Widen the fleet wanted-coord delta to int64 so it cannot overflow
UpdateFleetAndFlagshipNavigation subtracted two grid coordinates in int32_t.
Both coordinates span the signed 32-bit range, so their separation can exceed
what int32_t holds and the subtraction can overflow, which is undefined
behavior inside the fixed-tick, CRC-covered simulation path. Both operands now
convert to int64_t before the subtraction, where every representable pair fits.

Every consumer of the two deltas is a sign or non-zero test, so the direction
chosen for any pair that does not overflow today is unchanged, as are the
random-draw count and order in the function.

Also records the follow-up Plan for the pre-existing over-target size of
Projects/BrokenEngineSandbox/Source/Frame/AGENTS.md, and completes the Plan
this change implemented.

+++ Projects/BrokenEngineSandbox/Source/Frame/Collections/Players/PlayersNavigation.cpp
@@ -67,8 +67,9 @@ void XM_CALLCONV UpdateFleetAndFlagshipNavigation(Frame& __restrict rFrame, Play
 	// Fleet navigation: navigate toward fleet's wanted coord after countdown expires
 	if (!(fleetWantedCoord == rStaticData.coord) && uiPendingFleetWantedCoordTicks == 0)
 	{
-		int32_t iDeltaX = fleetWantedCoord.x - rStaticData.coord.x;
-		int32_t iDeltaY = fleetWantedCoord.y - rStaticData.coord.y;
+		// Two coords anywhere in the signed-int32 identity range can separate by more than int32 holds, so each converts to int64 ahead of the subtraction.
+		int64_t iDeltaX = static_cast<int64_t>(fleetWantedCoord.x) - static_cast<int64_t>(rStaticData.coord.x);
+		int64_t iDeltaY = static_cast<int64_t>(fleetWantedCoord.y) - static_cast<int64_t>(rStaticData.coord.y);
 
 		// Check if already heading in a valid direction toward wanted coord
 		bool bAlreadyValid = false;
